Refactor: Event directory
