+++ green/src/main/java/com/greengrim/green/core/event/controller/GetEventController.java
@@ -1,5 +1,6 @@
-package com.greengrim.green.core.event;
+package com.greengrim.green.core.event.controller;
 
+import com.greengrim.green.core.event.service.GetEventService;
 import com.greengrim.green.core.event.dto.EventResponseDto.EventInfo;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -13,9 +14,9 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 @RequestMapping("/visitor/events")
 @Tag(name = "이벤트")
-public class EventController {
+public class GetEventController {
 
-    private final EventService eventService;
+    private final GetEventService getEventService;
 
     /**
      * [GET] 홈 화면 이벤트 조회
@@ -23,6 +24,6 @@ public class EventController {
     @Operation(summary = "홈 화면 이벤트 조회")
     @GetMapping("/home")
     public ResponseEntity<EventInfo> getHomeEvent() {
-        return ResponseEntity.ok(eventService.getRecentEvent());
+        return ResponseEntity.ok(getEventService.getRecentEvent());
     }
 }
